wip: Handle other types of formulas, reuse same component for TopicBasedFormulas
