+++ src/containers/SetZoneTopicContainer.js
@@ -1,16 +1,16 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-import TopicBasedFormulas from "../../components/Formula/TopicBasedFormulas";
+import TopicBasedFormulas from "../components/Formula/TopicBasedFormulas";
 
 const SetChildrenContainer = props => {
-  const { topics, multiEntitiesFormulas } = props.set;
+  const { topics, zoneTopicFormulas } = props.set;
 
   return (
     <TopicBasedFormulas
       topics={topics}
       inputs={[]}
-      formulas={multiEntitiesFormulas}
+      formulas={zoneTopicFormulas}
     ></TopicBasedFormulas>
   );
 };
